feat(config): modify config module env var override
This feature modifies the way how to config module overrides the configuration using env vars. In
the previous config module in order to override the configuration you should create a env var like:
d4n.database.url. As creating env vars with dots is not allowed in some operative systems, we modify
that behaviour, and now the way to override is: DATABASE_URL

BREAKING CHANGE: Prefix is not longer necessary. The environment variables to override the values
now have a different format